Ensure that Vector component IDs are globally unique.
Vector requires that all component IDs do not have conflicting names,
regardless of their type. This is likely to ensure that internal telemetry
does not end up with label values that conflict with each other.

Bootstrapped some unit tests to cover this, and to ensure the source
project filter is always present.